requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.8
Pillow==10.0.0
protobuf==4.21.12
PyMySQL==1.0.3
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
QScintilla==2.14.1
requests==2.31.0
SQLAlchemy==2.0.15
tqdm==4.65.0
typing_extensions==4.6.1
urllib3==2.0.3
yarl==1.9.2
